requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Hugging Face ecosystem
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
datasets>=2.10.0

# Image and video processing
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Face detection and processing
face-recognition>=1.3.0
mediapipe>=0.10.0
insightface>=0.7.3

# API and web framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
aiohttp>=3.8.0

# RunPod specific
runpod>=1.5.0

# Optional GPU acceleration
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# xformers>=0.0.20  # For memory efficient attention

# Model-specific dependencies
# safetensors>=0.3.0
# controlnet-aux>=0.4.0